+++ backend/src/deck/routes.py
@@ -23,6 +23,8 @@
 '''routes.py is a file in deck folder that has all the functions defined that manipulate the deck. All CRUD functions are defined here.'''
 from flask import Blueprint, jsonify, request
 from flask_cors import cross_origin
+from datetime import datetime
+
 try:
     from .. import firebase
 except ImportError:
@@ -172,3 +174,101 @@ def delete(id):
             message = f'Delete Deck Failed {e}',
             status = 400
         ), 400
+
+
+
+@deck_bp.route('/deck/<deckId>/leaderboard', methods=['GET'])
+@cross_origin(supports_credentials=True)
+def get_leaderboard(deckId):
+    '''This endpoint fetches the leaderboard data for a specific deck.'''
+    try:
+        # Fetch leaderboard data for the given deck
+        leaderboard_entries = db.child("leaderboard").child(deckId).get()
+        leaderboard = []
+        for entry in leaderboard_entries.each():
+            data = entry.val()
+            leaderboard.append({
+                "userEmail": data.get("userEmail"),
+                "correct": data.get("correct", 0),
+                "incorrect": data.get("incorrect", 0),
+                "lastAttempt": data.get("lastAttempt")
+            })
+
+        # Sort leaderboard by score (correct answers) then by last attempt (descending)
+        leaderboard.sort(key=lambda x: (x["correct"], x["lastAttempt"]), reverse=True)
+
+        return jsonify({
+            "leaderboard": leaderboard,
+            "message": "Leaderboard data fetched successfully",
+            "status": 200
+        }), 200
+    except Exception as e:
+        return jsonify({
+            "leaderboard": [],
+            "message": f"An error occurred: {e}",
+            "status": 400
+        }), 400
+    
+@deck_bp.route('/deck/<deck_id>/update-leaderboard', methods=['POST'])
+@cross_origin(supports_credentials=True)
+def update_leaderboard(deck_id):
+    try:
+        data = request.get_json()
+        # Extract values from the request body
+        user_id = data.get("userId")  # Get userId from request body
+        user_email = data.get("userEmail")  # Keep for logging or notification
+        correct = data.get("correct")
+        incorrect = data.get("incorrect")
+
+        if not user_id:
+            return jsonify({"message": "User ID is required"}), 400  # Validate userId presence
+
+        # Use user_id from request body to update the leaderboard
+        leaderboard_ref = db.child("leaderboard").child(deck_id).child(user_id)
+        leaderboard_ref.update({
+            "userEmail": user_email,
+            "correct": correct,
+            "incorrect": incorrect,
+            "lastAttempt": datetime.now().isoformat()
+        })
+
+        return jsonify({"message": "Leaderboard updated successfully"}), 200
+
+    except Exception as e:
+        return jsonify({"message": "Failed to update leaderboard", "error": str(e)}), 500
+    
+@deck_bp.route('/deck/<deckId>/user-score/<userId>', methods=['GET'])
+@cross_origin(supports_credentials=True)
+def get_user_score(deckId, userId):
+    '''This endpoint fetches the user's score for a specific deck. If the user doesn't exist, return zero for all score values.'''
+    try:
+        # Fetch the user's leaderboard entry for the specified deck
+        leaderboard_entry = db.child("leaderboard").child(deckId).child(userId).get()
+
+        if leaderboard_entry.val() is not None:  # Check if the entry has data
+            data = leaderboard_entry.val()  # Get the value of the entry
+            score_data = {
+                "correct": data.get("correct", 0),
+                "incorrect": data.get("incorrect", 0),
+            }
+            return jsonify({
+                "score": score_data,
+                "message": "User score fetched successfully",
+                "status": 200
+            }), 200
+        else:
+            # Return zero for all score values if no entry exists
+            return jsonify({
+                "score": {
+                    "correct": 0,
+                    "incorrect": 0
+                },
+                "message": "No score found for the user, returning zeros.",
+                "status": 200  # Not Found status, as the user has no scores yet
+            }), 200
+
+    except Exception as e:
+        return jsonify({
+            "message": f"An error occurred: {e}",
+            "status": 400
+        }), 400
